requirements.txt
Django==5.0.0
psycopg[binary]==3.1.18
python-dotenv==1.0.1
django-split-settings==1.3.2
django-debug-toolbar==5.0.1
ipython==8.32.0
django-cors-headers==4.7.0
gunicorn==23.0.0
black==25.1.0
isort==6.0.1
redis==5.2.1
pydantic==2.10.6